Refactored Plato files for separate Translate and BoolToConcept executables
