+++ plato.cabal
@@ -27,9 +27,9 @@ library
                       Tuura.Concept.FSM.Simulation,
                       Tuura.Concept.FSM.Translation,
                       Tuura.Parser.Boolean,
-                      Tuura.Plato.Options,
-                      Tuura.Plato.Translation,
-                      Tuura.Plato.BooleanFunctions
+                      Tuura.Plato.Translate.Options,
+                      Tuura.Plato.Translate.Translation,
+                      Tuura.Plato.BoolToConcept.BooleanFunctions
     other-modules:    Tuura.Concept.Abstract,
                       Tuura.Concept.Circuit,
                       Tuura.Concept.Circuit.Basic,
