+++ src/Tuura/Plato/Translate/Options.hs
@@ -1,4 +1,4 @@
-module Tuura.Plato.Options (Options(..), getOptions) where
+module Tuura.Plato.Translate.Options (Options(..), getOptions) where
 
 import Data.Foldable (foldlM)
 import System.Console.GetOpt
@@ -9,7 +9,6 @@ data Options = Options
     , optOutput  :: String -> IO ()
     , optInclude :: [String]
     , optFSM     :: Bool
-    , optBool    :: Bool
     , optHelp    :: Bool }
 
 defaultOptions :: Options
@@ -18,7 +17,6 @@ defaultOptions   = Options
     , optOutput  = putStrLn
     , optInclude = []
     , optFSM     = False
-    , optBool    = False
     , optHelp    = False }
 
 options :: [OptDescr (Options -> IO Options)]
@@ -33,9 +31,6 @@ options =
  , Option ['f'] ["fsm"]
      (NoArg (\ opts -> return opts { optFSM = True }))
      "Translate concept specification to an FSM"
- , Option ['b'] ["boolean"]
-     (NoArg (\ opts -> return opts { optBool = True }))
-     "Create concept specification from Boolean set and reset functions"
  , Option ['h'] ["help"]
      (NoArg (\ opts -> return opts { optHelp = True }))
      "Show this help message"
@@ -50,34 +45,10 @@ getOptions = do
       (_, [] , _   ) -> ioError (userError ("\nNo input file given\n" ++ helpMessage))
       (_, _  , []  ) -> ioError (userError ("\nToo many input files\n" ++ helpMessage))
       (_, _  , errs) -> ioError (userError (concat errs ++ helpMessage))
-   _ <- validateOptions result
    return $ result
-
-    -- result <- case getOpt Permute options argv of
-    --     (opts, [] , []  ) -> foldlM (flip id) defaultOptions opts
-    --     (opts, [f], []  ) -> foldlM (flip id)
-    --                          defaultOptions { optInput = readFile f } opts
-    --     (_   , _  , []  ) -> ioError $ userError "Multiple input files"
-    --     (_   , _  , errs) -> ioError . userError $ concat errs
 
 helpMessage :: String
 helpMessage = usageInfo header options
   where
     header = "Usage: " ++ runCommand ++ " [input file] [OPTION...]"
     runCommand = "stack runghc translate/Main.hs --"
-
-validateOptions :: Options -> IO Options
-validateOptions ops
-    | (optInput ops /= "") && (optBool ops)   = ioError (userError
-      ("\nCannot translate a specification and generate a specification\n" ++
-      "from set and reset functions at the same time.\n" ++
-      "Remove the input filepath\n." ++ helpMessage))
-    | (optInclude ops /= []) && (optBool ops) = ioError (userError
-      ("\nIncludes cannot be used when generating a specification from\n" ++
-      "from set and reset functions at the same time\n" ++
-      "Remove the include flag and filepath.\n" ++ helpMessage))
-    | (optFSM ops) && (optBool ops)           = ioError (userError
-      ("\nGenerating a concept specification from set and reset functions\n" ++
-       "does not translate to FSM or STG specifically.\n" ++
-       "Remove the FSM translation flag.\n" ++ helpMessage))
-    | otherwise = return $ ops
